fix(local-web): move ReleaseNotesDialog into _app and add ExecutionProcessesProvider to vscode routes
- Move ReleaseNotesDialog useEffect from __root.tsx into _app.tsx as
  ReleaseNotesHandler component, placed inside NiceModalProvider to
  prevent 'No dispatch method detected' error on root route load
- Add missing ExecutionProcessesProviderWrapper to both standalone
  VSCode routes to fix useExecutionProcessesContext crash in
  ConversationList/RetryUiProvider
