+++ packages/local-web/src/routes/__root.tsx
@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import { Outlet, createRootRoute, useLocation } from '@tanstack/react-router';
+import { Outlet, createRootRoute } from '@tanstack/react-router';
 import { I18nextProvider } from 'react-i18next';
 import { usePostHog } from 'posthog-js/react';
 import { ThemeMode } from 'shared/types';
@@ -7,14 +7,12 @@ import i18n from '@/i18n';
 import { useUserSystem } from '@/shared/hooks/useUserSystem';
 import { ThemeProvider } from '@web/app/providers/ThemeProvider';
 import { useUiPreferencesScratch } from '@/shared/hooks/useUiPreferencesScratch';
-import { ReleaseNotesDialog } from '@/shared/dialogs/global/ReleaseNotesDialog';
 import { UserProvider } from '@/shared/providers/remote/UserProvider';
 import '@/app/styles/new/index.css';
 
 function RootRouteComponent() {
-  const { config, machineId, updateAndSaveConfig } = useUserSystem();
+  const { config, machineId } = useUserSystem();
   const posthog = usePostHog();
-  const location = useLocation();
 
   useUiPreferencesScratch();
 
@@ -30,33 +28,6 @@ function RootRouteComponent() {
       console.log('[Analytics] Analytics disabled by user preference');
     }
   }, [config?.analytics_enabled, machineId, posthog]);
-
-  useEffect(() => {
-    if (!config || !config.remote_onboarding_acknowledged) return;
-
-    const pathname = location.pathname;
-    if (pathname.startsWith('/onboarding') || pathname.startsWith('/migrate')) {
-      return;
-    }
-
-    let cancelled = false;
-
-    const showReleaseNotes = async () => {
-      if (config.show_release_notes) {
-        await ReleaseNotesDialog.show();
-        if (!cancelled) {
-          await updateAndSaveConfig({ show_release_notes: false });
-        }
-        ReleaseNotesDialog.hide();
-      }
-    };
-
-    void showReleaseNotes();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [config, updateAndSaveConfig, location.pathname]);
 
   return (
     <I18nextProvider i18n={i18n}>
