+++ packages/local-web/src/routes/_app.tsx
@@ -1,5 +1,9 @@
-import { type ReactNode } from 'react';
-import { createFileRoute, useParams } from '@tanstack/react-router';
+import { useEffect, type ReactNode } from 'react';
+import {
+  createFileRoute,
+  useParams,
+  useLocation,
+} from '@tanstack/react-router';
 import { Provider as NiceModalProvider } from '@ebay/nice-modal-react';
 import { SequenceTrackerProvider } from '@/shared/keyboard/SequenceTracker';
 import { SequenceIndicator } from '@/shared/keyboard/SequenceIndicator';
@@ -7,6 +11,7 @@ import { useWorkspaceShortcuts } from '@/shared/keyboard/useWorkspaceShortcuts';
 import { useIssueShortcuts } from '@/shared/keyboard/useIssueShortcuts';
 import { useKeyShowHelp, Scope } from '@/shared/keyboard';
 import { KeyboardShortcutsDialog } from '@/shared/dialogs/shared/KeyboardShortcutsDialog';
+import { ReleaseNotesDialog } from '@/shared/dialogs/global/ReleaseNotesDialog';
 import { TerminalProvider } from '@/shared/providers/TerminalProvider';
 import { HostIdProvider } from '@/shared/providers/HostIdProvider';
 import { WorkspaceProvider } from '@/shared/providers/WorkspaceProvider';
@@ -14,6 +19,7 @@ import { ExecutionProcessesProvider } from '@/shared/providers/ExecutionProcesse
 import { LogsPanelProvider } from '@/shared/providers/LogsPanelProvider';
 import { ActionsProvider } from '@/shared/providers/ActionsProvider';
 import { useWorkspaceContext } from '@/shared/hooks/useWorkspaceContext';
+import { useUserSystem } from '@/shared/hooks/useUserSystem';
 import { SharedAppLayout } from '@/shared/components/ui-new/containers/SharedAppLayout';
 
 function KeyboardShortcutsHandler() {
@@ -25,6 +31,40 @@ function KeyboardShortcutsHandler() {
   );
   useWorkspaceShortcuts();
   useIssueShortcuts();
+  return null;
+}
+
+function ReleaseNotesHandler() {
+  const { config, updateAndSaveConfig } = useUserSystem();
+  const location = useLocation();
+
+  useEffect(() => {
+    if (!config || !config.remote_onboarding_acknowledged) return;
+
+    const pathname = location.pathname;
+    if (pathname.startsWith('/onboarding') || pathname.startsWith('/migrate')) {
+      return;
+    }
+
+    let cancelled = false;
+
+    const showReleaseNotes = async () => {
+      if (config.show_release_notes) {
+        await ReleaseNotesDialog.show();
+        if (!cancelled) {
+          await updateAndSaveConfig({ show_release_notes: false });
+        }
+        ReleaseNotesDialog.hide();
+      }
+    };
+
+    void showReleaseNotes();
+
+    return () => {
+      cancelled = true;
+    };
+  }, [config, updateAndSaveConfig, location.pathname]);
+
   return null;
 }
 
@@ -66,6 +106,7 @@ function AppLayoutRouteComponent() {
 
   return (
     <AppRouteProviders key={hostId ?? 'local'}>
+      <ReleaseNotesHandler />
       <SequenceTrackerProvider>
         <SequenceIndicator />
         <KeyboardShortcutsHandler />
